Add files via upload
The program assumes that the input list forms a palindrome. In a palindrome, the middle characters would be the same from both ends. Therefore, to find the middle character, we can consider the last character of the palindrome.
The Middle_Character function takes a list as input and returns the last element of the list, which represents the middle character of the assumed palindrome.
Example:
For the input list ['a', 'm', 'm', 'a', 'd'], the function returns 'd' because it is the last character of the list and represents the middle character of the palindrome formed by the list.
For the input list ['i', 'i', 'e', 'd', 'e', 'r', 'd', 'r', 'v'], the function returns 'v', which is the last character of the list and represents the middle character of the palindrome formed by the list.