Add files via upload
We have added the sorted function to sort the elements in the resulting lists for better readability.
The duplicates function takes a list as input and separates it into two lists: duplicates and unique.
First, we create a dictionary called frequency to keep track of the occurrence of each element in the list.
Next, we iterate over the frequency dictionary to separate the elements into the duplicates and unique lists. If the count for an element is greater than 1, it means it has duplicates, so we add it to the duplicates list. Otherwise, if the count is equal to 1, we add it to the unique list. Finally, we return the duplicates and unique lists.


In the example, we provide the list names a [1, 5, 1, 3, 7, 21, 25, 2, 5, 3, 21] to the duplicates function. We then print the resulting duplicates and unique lists, which are [1, 3, 5, 21] and [2, 7, 25], respectively.